Rename Model directories to Models and update namespaces
