+++ hugashop/crm/Extensions/FacebookCommerce/Models/FeedGenerator.php
@@ -44,7 +44,7 @@
  * 
  */
 
-namespace HugaShop\Extensions\FacebookCommerce\Model;
+namespace HugaShop\Extensions\FacebookCommerce\Models;
 
 use HugaShop\Models\Image;
 use HugaShop\Models\Config;
