+++ hugashop/crm/Extensions/InfoBlock/Models/InfoBlock.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace HugaShop\Extensions\InfoBlock\Model;
+namespace HugaShop\Extensions\InfoBlock\Models;
 
 use HugaShop\Extensions\BaseExtensionModel;
