+++ hugashop/crm/Extensions/SmsSender/Models/SmsSenderMail.php
@@ -8,7 +8,7 @@
  *
  */
 
-namespace HugaShop\Extensions\SmsSender\Model;
+namespace HugaShop\Extensions\SmsSender\Models;
 
 use HugaShop\Models\User\UserMailing;
 use HugaShop\Extensions\BaseExtensionModel;
